% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_download.R
\name{gbif_download}
\alias{gbif_download}
\title{Download GBIF data using aws.s3 sync}
\usage{
gbif_download(version = "2021-11-01", dir = gbif_dir())
}
\arguments{
\item{version}{'prefix' string (folder) from the
https://registry.opendata.aws/gbif/ which should be synced.}

\item{dir}{path to local directory where parquet files should be stored}
}
\description{
Sync a local directory with selected release of the AWS copy of GBIF
}
\details{
Sync parquet files from GBIF public data catalogue,
https://registry.opendata.aws/gbif/

Note that data can also be found on the Microsoft Cloud,
https://planetarycomputer.microsoft.com/dataset/gbif

Also, some users may prefer to download this data using an alternative
interface or work on a cloud-host machine where data is already available.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gbif_download()
\dontshow{\}) # examplesIf}
}
