% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_example_data.R
\name{gbif_example_data}
\alias{gbif_example_data}
\title{Return a path to the directory containing GBIF example parquet data}
\usage{
gbif_example_data()
}
\value{
path to the example occurrence data installed with the package.
}
\description{
Return a path to the directory containing GBIF example parquet data
}
\details{
example data is taken from the first 1000 rows of the
2011-11-01 release of the parquet data.
}
\examples{
gbif_example_data()
}
