% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_remote.R
\name{gbif_remote}
\alias{gbif_remote}
\title{gbif remote}
\usage{
gbif_remote(
  version = gbif_version(),
  bucket = gbif_default_bucket(),
  to_duckdb = FALSE,
  safe = TRUE,
  ...
)
}
\arguments{
\item{version}{GBIF snapshot date}

\item{bucket}{GBIF bucket name (including region)}

\item{to_duckdb}{Return a remote duckdb connection or arrow connection?}

\item{safe}{logical, default TRUE.  Should we exclude columns \code{mediatype} and \code{issue}?
varchar datatype on these columns substantially slows downs queries.}

\item{...}{additional parameters passed to the s3_bucket() (e.g. for remote
access to independently hosted buckets)}
}
\value{
a remote tibble \code{tbl_sql} class object (by default), or a arrow
Dataset query if \code{to_duckdb} is FALSE.  In either case, users should call
\verb{[dplyr::collect]} on the final result to force evaluation and bring the
resulting data into
memory in R.
}
\description{
Connect to GBIF remote directly. Can be much faster than downloading
for one-off use or when using the package from a server in the same region
as the data. See Details.
}
\details{
Query performance is dramatically improved in queries that return only
a subset of columns. Consider using explicit \code{select()} commands to return only
the columns you need.

A summary of this GBIF data, along with column meanings can be found at
\url{https://github.com/gbif/occurrence/blob/master/aws-public-data.md}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

gbif <- gbif_remote()
gbif()
\dontshow{\}) # examplesIf}
}
